refactor: NumberPicker 숫자 지정하는 기능을 프래그먼트로 이동
