+++ android/app/src/main/java/com/woowacourse/ody/presentation/time/MeetingTimeFragment.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.NumberPicker
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.woowacourse.ody.databinding.FragmentMeetingTimeBinding
@@ -39,12 +40,30 @@ class MeetingTimeFragment : Fragment() {
     }
 
     private fun initializeView() {
-        binding.npMeetingTimeHour.wrapSelectorWheel = true
-        binding.npMeetingTimeMinute.wrapSelectorWheel = true
+        binding.npMeetingTimeHour.setRangeValues(MEETING_HOUR_RANGE)
+        binding.npMeetingTimeMinute.setRangeValues(MEETING_MINUTE_RANGE)
+    }
+
+    private fun NumberPicker.setRangeValues(values: IntRange) {
+        wrapSelectorWheel = true
+        val displayValues = values.toStrings()
+        minValue = 0
+        maxValue = displayValues.size - 1
+        displayedValues = displayValues.toTypedArray()
+    }
+
+    private fun IntRange.toStrings() = map {
+        val string = it.toString()
+        if (string.length == 1) "0$string" else string
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
+    }
+
+    companion object {
+        private val MEETING_HOUR_RANGE = 0..23
+        private val MEETING_MINUTE_RANGE = 0..59
     }
 }
